ingestion/requirements.txt
# NBA data ingestion — Python 3.10+
nba_api>=1.4
psycopg2-binary>=2.9
python-dotenv>=1.0
